rename modules with long names
