+++ services/director-v2/src/simcore_service_director_v2/core/application.py
@@ -25,7 +25,7 @@ from ..modules import (
     dynamic_services,
     dynamic_sidecar,
     node_rights,
-    o2vars_substitutions,
+    osparc_variables_substitutions,
     rabbitmq,
     remote_debug,
     storage,
@@ -132,7 +132,7 @@ def init_app(settings: AppSettings | None = None) -> FastAPI:
         settings = app.state.settings
     assert settings  # nosec
 
-    o2vars_substitutions.setup(app)
+    osparc_variables_substitutions.setup(app)
 
     if settings.SC_BOOT_MODE == BootModeEnum.DEBUG:
         remote_debug.setup(app)
