+++ services/director-v2/src/simcore_service_director_v2/utils/osparc_session_variables.py
